requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itypes==1.2.0
Jinja2==3.1.4
joblib==1.4.2
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.2.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.2
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.0
transformers==4.46.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0
zope.interface==7.2
